requirements.txt
streamlit
pandas
numpy
datetime
scikit-learn
keras.models
yfinance
tensorflow
requests
matplotlib
timedelta
keras.layers
keras.initializers


